+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Changes
 
+- ListMenu.activeItem accepts null to signal that no menu item is active
 - Removed `children` parameter from `hbox` and `vbox` builders - they were early remnants from before we realized how powerful builders could be :)
 - `action` delegate no longer has `ActionEvent` as parameter so it can be used for no-args function references. Fallback to `setOnAction` if you need the event.